changed hexbinary handling to use asserticals register_value_generator to change all strings to hexbinary8. Necessary as child class handling with hexbinary generation proved challenging. Some classes are manually skipped now (e.g. ConnectionPointRequest) due to intentional changes from the schema, and will be added as individual tests later. The type for DERmodessupported and DERmodesenabled have been updated to hexbinary8 to conform with schema.

+++ src/envoy_schema/server/schema/sep2/event.py
@@ -4,7 +4,7 @@ from pydantic_xml import element
 
 from envoy_schema.server.schema.sep2.identification import RespondableSubscribableIdentifiedObject
 from envoy_schema.server.schema.sep2.types import DateTimeIntervalType, OneHourRangeType, TimeType
-from envoy_schema.server.schema.sep2.primitive_types import String192, Uint8
+from envoy_schema.server.schema.sep2.primitive_types import String192
 from envoy_schema.server.schema.sep2.base import BaseXmlModelWithNS
 
 
@@ -14,7 +14,7 @@ class EventStatus(BaseXmlModelWithNS):
     the event.  Devices can also subscribe to a specific resource instance to get updates when any of its attributes
     change, including the Status object."""
 
-    currentStatus: Uint8 = element()
+    currentStatus: int = element()
     dateTime: TimeType = element()
     potentiallySuperseded: bool = element()
     potentiallySupersededTime: Optional[TimeType] = element()
